# komplett-server-dev/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Node-Typescript API",
    "version": "1.0.0",
    "description": "A sample API"
  },
  "servers": [
    {
      "url": "http://192.168.0.119:3000"
    },
    {
      "url": "https://mongo-kompletee.herokuapp.com"
    },
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "http://ec2-18-216-130-99.us-east-2.compute.amazonaws.com:3000"
    },
    {
      "url": "http://ec2-3-22-60-204.us-east-2.compute.amazonaws.com:3000"
    }
  ],
  "paths": {
...
    "/api/master-materials": {
      "get": {
        "description": "Get master material api with pagination and sorting",
        "tags": [
          "MasterMaterial"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "schema": {
              "type": "string"
            },
            "description": "Carpentry"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": 10
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Example = createdAt"
          },
          {
            "in": "query",
            "name": "order",
            "schema": {
              "type": "acs"
            },
            "description": "Example = acs"
          }
        ],
        "responses": {
          "200": {
            "description": "returns customers list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SubMaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "add master sub materials",
        "tags": [
          "MasterMaterial"
        ],
        "requestBody": {
          "description": "application and category field is required",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubMaterialSchema"
              },
              "example": {
                "application": "application name here",
                "category": "Engineering"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/InsertSubMaterial"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "description": "update master-materials",
        "tags": [
          "MasterMaterial"
        ],
        "requestBody": {
          "description": "_id field is required",
          "required": true,
          "content": {
            "application/json": {
              "required": [
                "id"
              ],
              "schema": {
                "$ref": "#/components/schemas/SubMaterialSchema"
              },
              "example": {
                "_id": "5f9bbb17405a9420e895920f",
                "application": "application name here",
                "category": "Engineering"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "return updated master-material",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SubMaterialSchema"
                    }
                  ]
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/master-materials/{id}": {
      "delete": {
        "description": "delete master-materials by id",
        "tags": [
          "MasterMaterial"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
    "/api/material": {
      "get": {
        "description": "Get Materials",
        "tags": [
          "material"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": 1
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": 10
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string"
            },
            "description": "Example = Engineering"
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string"
            },
            "description": "Example = _id"
          }
        ],
        "responses": {
          "200": {
            "description": "return materials based on conditions",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "add material",
        "tags": [
          "material"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSchema"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "material created",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "material"
        ],
        "requestBody": {
          "required": true,
          "description": "_id field is required",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaterialSchema"
              },
              "example": {
                "_id": "5f85845341855c2b209ba312",
                "type": "Engineering",
                "subMaterials": []
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/MaterialSchema"
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/material/{id}": {
      "delete": {
        "description": "delete material by id",
        "tags": [
          "material"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "content": null
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MaterialSchema": {
        "properties": {
          "crowd": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "buildingComponents": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "level": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "subMaterials": {
            "type": "array",
            "example": [],
            "items": {
              "$ref": "#/components/schemas/SubMaterialSchema"
            }
          }
        }
      },
      "SubMaterialSchema": {
        "properties": {
          "subject": {
            "type": "string"
          },
          "NS3420": {
            "type": "string"
          },
          "application": {
            "type": "string"
          },
          "crowd": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "itemPrice": {
            "type": "number"
          },
          "itemFactor": {
            "type": "number"
          },
          "inverseFactor": {
            "type": "number"
          },
          "useListPrice": {
            "type": "number"
          },
          "documents": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "quantity": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string"
              },
              "quantityPerComponent": {
                "type": "number"
              },
              "svinn": {
                "type": "number"
              },
              "inTotal": {
                "type": "number"
              }
            }
          },
          "time": {
            "type": "object",
            "properties": {
              "minPerComponent": {
                "type": "number"
              },
              "hoursPerComponent": {
                "type": "number"
              },
              "totalTimeConsumption": {
                "type": "string"
              }
            }
          }
        }
      },
      "InsertSubMaterial": {
        "allOf": [
          {
            "$ref": "#/components/schemas/SubMaterialSchema"
          },
          {
            "type": "object",
            "properties": {
              "materialId": {
                "type": "string"
              }
            }
          }
        ]
      },
...
      "Error": {
        "type": "object",
        "required": [
          "status",
          "message"
        ],
        "properties": {
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Error description",
            "example": "User created"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-access-token"
      }
    }
  },
...
}